requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
discord==1.7.3
discord.py==1.7.3
Flask==2.0.2
google-trans-new==1.1.9
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
Pillow==9.0.0
praw==7.5.0
prawcore==2.3.0
pycodestyle==2.8.0
python-dotenv==0.19.2
replit==3.2.4
requests==2.27.1
soupsieve==2.3.1
toml==0.10.2
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.2
yarl==1.7.2
